Enable Function subscription naming workaround when activated
The Pulsar Function will use the function name as the subscription name when this workaround
is activated. This workaround is activated when -DdsPatchFunctionNameAsSubscriptionName=true
system property or dsPatchFunctionNameAsSubscriptionName=true environment variable is passed
to the Functions Worker (or to the Pulsar Broker when there isn't a separate Functions Worker).

(cherry picked from commit b2e95a6457278943eca47193595d112cd680348b)