Close the replicator and replication client when delete cluster. (#11342)
Currently, when a cluster has been deleted, the replicator and the replication client will not be closed.
The producer of the replicator will try to reconnect to the deleted cluster continuously.
We should close and clean up the replicator and the replication client for the deleted cluster.

A new test was added for verifying the replicator and the replication client will be closed after the cluster is deleted.

(cherry picked from commit 8a4147e3373da00cd11bd81416d2afbed00104b7)