Function: add possibility to pass additional JVM arguments to the function JVM (additionalJavaRuntimeArguments) (#13282)
### Motivation

Sometimes it would be useful to be able to tune the JVM of the functions by passing additional command line arguments,
like "-XX:+ExitOnOutOfMemoryError" or "-Dlog4j2.formatMsgNoLookups".

Those settings are not per-function, but they are to be applied to every function process.

### Modifications

Add a new configuration parameter additionalJavaRuntimeArguments in functions_worker.conf.
This is a list of strings to be added to the command line of the Java process.

(cherry picked from commit 58e1dff4e83ec95d9f0805b1072f34bf5250e989)